Use Options instead of customProps
the renderNodes functios now takes a RenderOptions object instead of a
props object. These options are passed to the renderers in the
constructor instead of in the renderNode function. These options allow
more configuration of rendering. The only supported options are
className and allowSoftBreaks.
line breaks are no longer rendered as br tags by default. This feature
is now hidden behind the allowSoftBreaks flag.